+++ modules/matrix-bridge/test/matrix-event-handler-test.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const assert = require('assert');
+const sinon = require('sinon');
 const fixtureLoader = require('gitter-web-test-utils/lib/test-fixtures');
 const mongoUtils = require('gitter-web-persistence-utils/lib/mongo-utils');
 const MatrixEventHandler = require('../lib/matrix-event-handler');
@@ -21,12 +22,14 @@ function createEventData(extraEventData) {
 
 describe('matrix-event-handler', () => {
   let matrixEventHandler;
-  const matrixBridge = {
-    getIntent: (/*userId*/) => ({
+  let matrixBridge;
+  beforeEach(() => {
+    const intentSpies = {
       createRoom: () => ({
         room_id: `!${fixtureLoader.generateGithubId()}:localhost`
       }),
       createAlias: () => {},
+      join: sinon.spy(),
       getProfileInfo: () => ({
         displayname: 'Alice',
         avatar_url: 'mxc://abcedf1234'
@@ -34,8 +37,12 @@ describe('matrix-event-handler', () => {
       getClient: (/*avatarUrl*/) => ({
         mxcUrlToHttp: () => 'myavatar.png'
       })
-    })
-  };
+    };
+
+    matrixBridge = {
+      getIntent: (/*userId*/) => intentSpies
+    };
+  });
 
   describe('onAliasQuery', () => {
     const fixture = fixtureLoader.setupEach({
@@ -139,6 +146,17 @@ describe('matrix-event-handler', () => {
           troupe: 'troupe1',
           text: 'my original message from matrix'
         },
+        messageOldFromVirtualUser1: {
+          user: 'userBridge1',
+          virtualUser: {
+            type: 'matrix',
+            externalId: 'test-person:matrix.org',
+            displayName: 'Tessa'
+          },
+          troupe: 'troupe1',
+          text: 'my original old message from matrix',
+          sent: new Date('2020-12-01T00:00:00.000Z')
+        },
         messageStatusFromVirtualUser1: {
           user: 'userBridge1',
           virtualUser: {
@@ -179,7 +197,7 @@ describe('matrix-event-handler', () => {
 
         const messages = await chatService.findChatMessagesForTroupe(fixture.troupe1.id);
         const targetMessage = messages.find(m => m.id === fixture.messageFromVirtualUser1.id);
-        assert.strictEqual(messages.length, 2);
+        assert.strictEqual(messages.length, 3);
         assert.strictEqual(targetMessage.text, 'my edited message from matrix');
       });
 
@@ -209,13 +227,42 @@ describe('matrix-event-handler', () => {
         await matrixEventHandler.onEventData(eventData);
 
         const messages = await chatService.findChatMessagesForTroupe(fixture.troupe1.id);
-        assert.strictEqual(messages.length, 2);
+        assert.strictEqual(messages.length, 3);
         const targetMessage = messages.find(m => m.id === fixture.messageStatusFromVirtualUser1.id);
         assert.strictEqual(
           targetMessage.text,
           '@alice:localhost my edited emote/status message from matrix'
         );
         assert.strictEqual(targetMessage.status, true);
+      });
+
+      it('When we receive message edit from Matrix outside of the Gitter edit time window, send a new Gitter message with reference to original message the Gitter room', async () => {
+        const matrixMessageEventId = `$${fixtureLoader.generateGithubId()}`;
+        const eventData = createEventData({
+          type: 'm.room.message',
+          content: {
+            body: '* my edited message from matrix',
+            'm.new_content': { body: 'my edited message from matrix', msgtype: 'm.text' },
+            'm.relates_to': {
+              event_id: matrixMessageEventId,
+              rel_type: 'm.replace'
+            }
+          }
+        });
+        await store.storeBridgedMessage(
+          fixture.messageOldFromVirtualUser1,
+          eventData.room_id,
+          matrixMessageEventId
+        );
+
+        await matrixEventHandler.onEventData(eventData);
+
+        const messages = await chatService.findChatMessagesForTroupe(fixture.troupe1.id);
+        const targetMessage = messages.find(m => m.id === fixture.messageOldFromVirtualUser1.id);
+        const newMessage = messages[3];
+        assert.strictEqual(messages.length, 4);
+        assert.strictEqual(targetMessage.text, 'my original old message from matrix');
+        assert.strictEqual(newMessage.text.split('): ')[1], 'my edited message from matrix');
       });
 
       it('Ignore message edit from Matrix when there is no matching message', async () => {
@@ -243,7 +290,7 @@ describe('matrix-event-handler', () => {
 
         const messages = await chatService.findChatMessagesForTroupe(fixture.troupe1.id);
         const targetMessage = messages.find(m => m.id === fixture.messageFromVirtualUser1.id);
-        assert.strictEqual(messages.length, 2);
+        assert.strictEqual(messages.length, 3);
         assert.strictEqual(targetMessage.text, 'my original message from matrix');
       });
     });
@@ -588,6 +635,51 @@ describe('matrix-event-handler', () => {
         const messages = await chatService.findChatMessagesForTroupe(fixture.troupe1.id);
         assert.strictEqual(messages.length, 1);
         assert.strictEqual(messages[0].text, 'my original message from matrix');
+      });
+    });
+
+    describe('handleBotInvitationEvent', () => {
+      const fixture = fixtureLoader.setupEach({
+        userBridge1: {},
+        troupe1: {},
+        messageFromVirtualUser1: {
+          user: 'userBridge1',
+          virtualUser: {
+            type: 'matrix',
+            externalId: 'test-person:matrix.org',
+            displayName: 'Tessa'
+          },
+          troupe: 'troupe1',
+          text: 'my original message from matrix'
+        }
+      });
+
+      beforeEach(() => {
+        matrixEventHandler = new MatrixEventHandler(matrixBridge, fixture.userBridge1.username);
+      });
+
+      it('When we receive a Matrix invite, the bridge bot user joins the room', async () => {
+        const eventData = createEventData({
+          type: 'm.room.member',
+          state_key: '@gitter-badger:my.matrix.host'
+        });
+
+        await matrixEventHandler.onEventData(eventData);
+
+        // Room is created for something that hasn't been bridged before
+        assert.strictEqual(matrixBridge.getIntent().join.callCount, 1);
+      });
+
+      it('When we receive a Matrix invite for another user, does not cause our bot to try to join the room', async () => {
+        const eventData = createEventData({
+          type: 'm.room.member',
+          state_key: '@some-random-user:my.matrix.host'
+        });
+
+        await matrixEventHandler.onEventData(eventData);
+
+        // Room is created for something that hasn't been bridged before
+        assert.strictEqual(matrixBridge.getIntent().join.callCount, 0);
       });
     });
   });